fix: Add IMMUTABLE mode support for uint64 getters + format warning
✅ Fixes:
- dap_json_object_get_uint64_ext() now supports IMMUTABLE mode
- Corrected uint64 storage pointer usage with DAP_JSON_FLAG_MALLOC
- Fixed format warning in stage2_ref.c (%zu -> %u)

Remaining: Other test failures (unicode, invalid JSON validation, etc)

+++ module/json/src/stage2/dap_json_stage2_ref.c
@@ -1499,8 +1499,8 @@ static int32_t s_parse_string(
     *a_out_end_offset = a_start + l_end_offset;
     
     debug_if(dap_json_get_debug(), L_DEBUG,
-             "Phase 2.0.3: String ref[%d] - offset=%u, length=%zu, escapes=%d",
-             ref_index, (uint32_t)data_offset, l_scanned_string.length, 
+             "Phase 2.0.3: String ref[%d] - offset=%u, length=%u, escapes=%d",
+             ref_index, (uint32_t)data_offset, (uint32_t)l_scanned_string.length, 
              l_scanned_string.needs_unescape);
     
     return ref_index;
